Re-structure
While figuring out IK rigs, i learned that my structure for the player and NPCs needed major rework as well as a better inventory system.
- Created basic item pickups
- Created several enums (Equipment class, weapon type, and weapon names for each type)
- Converted NPC to use Manny SKM and animation blueprint
- Build the basics of a Base Character class. Will be used as the parent class for all players and NPCS
- Implemented basic animations for rifle and pistols

Still not at all done with this. Been too distracted to get it all done. Also restructured the inventory based on a video tutorial i found when I was trying to figure out IK rigging. Was a 2 hour video, but also had to figure out to take what they were doing and adjust it to my system (which is more than just guns).

Up next: Continuing on with the restructuring and finally implement player and the NPC to work based on this base character class. Left off with working on the animation blueprint for manny where i had empty state machines and caches acting in place for other weapons.  Plan is to use them to figure out the proper blend structure of the blueprint to get the right poses for the npc based off of the character's locomotive state and currently equipped item.