Basic NPC Standing animations and LHIK integration
Based on my limited testing (had to prevent npc from moving around), but basic spawning of npc with either an M1911 or an MP-40 shows that basic integration of animations and LHIK is complete. The blending is ok, but its as good as I can get with the resources I have.

Next: Need to figure out how to get the player character (the arms) to also mimic the animations as well as switching weapons. This will show if the NPC can fully do it to as well as see what needs fixing. (I know crouching will be an issue)