Project Cleanup
Reorganized the folders of the project and deleted old assets. Project is organized more or less by broad functionality. For instance, in the gameplay folder there are many things that can be interacted with or controlled, like characters for example. In that folder, I have things like the player vs npc and in their relevant folders are the animations, blueprints, meshes, textures, etc.